\name{barnacles1}
\alias{barnacles1}
\docType{data}
\title{
Sales of barnacles in Cedeira
}
\description{
Information about sales and prices of barnacles in two galician towns for each month from 2004 to 2013. The data have been transformed using the logarithm function.
}
\usage{data(barnacles1)}
\format{
A matrix containing 3 columns: 

\code{barnacles1[, 1]} contains the number of sales (in kg) of barnacles in Cedeira's fish market;

\code{barnacles1[, 2]} contains the prices (in euro/kg) of the barnacles in Cedeira's fish market;

\code{barnacles1[, 3]} contains the number of sales (in kg) of barnacles in Carino's fish market.
}
\source{
\url{http://dm.udc.es/modes/sites/default/files/barnacles1.rar}
\url{http://dm.udc.es/modes/sites/default/files/barnacles1.zip}
}
\keyword{datasets}