\name{barnacles2}
\alias{barnacles2}
\docType{data}
\title{
Sales of barnacles in Cangas
}
\description{
Information about sales and prices of barnacles in two galician towns for each month from 2004 to 2013. The data have been transformed using the logarithm function.
}
\usage{data(barnacles2)}
\format{
A matrix containing 3 columns: 

\code{barnacles1[, 1]} contains the number of sales (in kg) of barnacles in Cangas' fish market;

\code{barnacles1[, 2]} contains the prices (in euro/kg) of the barnacles in Cangas' fish market;

\code{barnacles1[, 3]} contains the number of sales (in kg) of barnacles in Baiona's fish market.
}
\source{
\url{http://dm.udc.es/modes/sites/default/files/barnacles2.rar} 
\url{http://dm.udc.es/modes/sites/default/files/barnacles2.zip}
}
\keyword{datasets}