\name{PLRModels-internal}
\alias{PLRModels-internal}
\alias{best.arima}
\alias{Epanechnikov}
\alias{gaussian}
\alias{quadratic}
\alias{symsolve}
\alias{triweight}
\alias{uniform}
\alias{var.cov.matrix}
\alias{var.cov.sum}
\title{Internal Functions}
\description{
  Internal functions
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
