\name{PLRModels-package}
\alias{PLRModels-package}
\alias{PLRModels}
\docType{package}
\title{
Statistical inference in partial linear regression models
}
\description{
This package provides statistical inference tools applied to Partial Linear Regression (PLR) models. Specifically, point estimation, confidence intervals estimation, bandwidth selection, goodness-of-fit tests and analysis of covariance are considered. Kernel-based methods, combined with ordinary least squares estimation, are used and time series errors are allowed. In addition, these techniques are also implemented for both parametric (linear) and nonparametric regression models.
}
\details{
\tabular{ll}{
Package: \tab PLRModels\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-01-01\cr
License: \tab GPL \cr
}
The most important functions are those directly related with the PLR models; that is, \code{plrm.gcv}, \code{plrm.cv}, \code{plrm.beta}, \code{plrm.est}, \code{plrm.gof}, \code{plrm.ancova} and \code{plrm.ci}. Although the other functions included in the package are auxiliary ones, they can be used independiently.
}
\author{
Authors: 
German Aneiros Perez <ganeiros@udc.es>

Ana Lopez Cheda <ana.lopez.cheda@udc.es>

Maintainer: German Aneiros Perez <ganeiros@udc.es>
}

\keyword{PLRModels}
